+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.16.0
+
+### Minor Changes
+
+-   4fb53ce: Fix sonar warnings
+
+## 0.15.0
+
+### Minor Changes
+
+-   e00e583: Controller Extension support
+
 ## 0.14.3
 
 ### Patch Changes