+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 0.14.0
+
+### Minor Changes
+
+-   4fb53ce: Fix sonar warnings
+
+### Patch Changes
+
+-   Updated dependencies [4fb53ce]
+    -   @sap-ux/ui5-application-writer@0.16.0
+
 ## 0.13.2
 
 ### Patch Changes